EUI-4615 New Task Check At Event Completion
