+++ src/index.ts
@@ -116,7 +116,6 @@ export { PaletteModule,
          ActivityModule,
          CaseHeaderModule,
          EventStartModule,
-         EventCompletionModule,
          EventLogModule,
          SearchFiltersModule,
          CaseResolver,
