+++ src/shared/components/case-editor/case-edit-submit/case-edit-submit.html
@@ -112,9 +112,9 @@
   </div>
   <p class="cancel"><a (click)="cancel()" href="javascript:void(0)" [class.disabled]="isSubmitting">{{getCancelText()}}</a></p>
 </form>
-<ccd-event-completion
+<ccd-case-event-completion
   [caseId]="getCaseId()"
   [eventId]="editForm.value.event.id"
   [task]="task"
   (eventCanBeCompleted)="onEventCanBeCompleted($event)">
-</ccd-event-completion>
+</ccd-case-event-completion>
