+++ src/shared/components/case-editor/case-event-completion/index.ts
@@ -1,0 +1,1 @@
+export * from './case-event-completion.component';
