+++ src/shared/components/case-editor/domain/event-completion-return-states.enum.model.ts
