+++ src/shared/components/case-editor/domain/event-completion-state-machine-context.model.ts
@@ -1,0 +1,14 @@
+import { ActivatedRoute, Router } from "@angular/router";
+import { Task } from "../../../domain/work-allocation/Task";
+import { SessionStorageService } from "../../../services/session/session-storage.service";
+import { WorkAllocationService } from "../services/work-allocation.service";
+
+export interface EventCompletionStateMachineContext {
+  task: Task;
+  caseId: string;
+  eventId: string;
+  router: Router;
+  route: ActivatedRoute;
+  sessionStorageService: SessionStorageService;
+  workAllocationService: WorkAllocationService;
+}
