+++ src/shared/components/case-editor/domain/event-completion-states.enum.model.ts
