+++ src/shared/components/case-editor/domain/index.ts
@@ -2,3 +2,6 @@ export * from './wizard.model';
 export * from './confirmation.model';
 export * from './wizard-page.model';
 export * from './wizard-page-field.model';
+export * from './event-completion-return-states.enum.model';
+export * from './event-completion-state-machine-context.model';
+export * from './event-completion-states.enum.model';
