+++ src/shared/components/case-viewer/viewer.routing.ts
@@ -37,16 +37,16 @@ export const viewerRouting: Routes = [
     canActivate: [EventStartGuard],
     canDeactivate: [FileUploadProgressGuard],
   },
-  {
-    path: 'trigger/:eid/task/:tid',
-    resolve: {
-      eventTrigger: EventTriggerResolver,
-    },
-    component: CaseEventTriggerComponent,
-    children: editorRouting,
-    canActivate: [EventStartGuard],
-    canDeactivate: [FileUploadProgressGuard],
-  },
+  // {
+  //   path: 'trigger/:eid/task/:tid',
+  //   resolve: {
+  //     eventTrigger: EventTriggerResolver,
+  //   },
+  //   component: CaseEventTriggerComponent,
+  //   children: editorRouting,
+  //   canActivate: [EventStartGuard],
+  //   canDeactivate: [FileUploadProgressGuard],
+  // },
   {
     path: 'event-start',
     component: EventStartComponent,
