--- src/shared/components/event-completion/index.ts
@@ -1,3 +1,0 @@
-export * from './event-completion.component';
-export * from './event-completion.module';
-export * from './services';
