--- src/shared/components/event-completion/models/index.ts
@@ -1,3 +1,0 @@
-export * from './event-completion-return-states.enum.model';
-export * from './event-completion-state-machine-context.model';
-export * from './event-completion-states.enum.model';
