--- src/shared/components/event-completion/services/index.ts
@@ -1,1 +1,0 @@
-export * from './event-completion-state-machine.service';
