+++ src/shared/components/event-start/event-start.module.ts
@@ -3,7 +3,6 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
 import { RouterModule } from '@angular/router';
-import { ActivityModule } from '../../components/activity';
 import { MultipleTasksExistComponent } from './components/multiple-tasks-exist/multiple-tasks-exist.component';
 import { NoTasksAvailableComponent } from './components/no-tasks-available/no-tasks-available.component';
 import { TaskAssignedComponent } from './components/task-assigned/task-assigned.component';
@@ -19,7 +18,6 @@ import { EventStartStateMachineService } from './services';
   imports: [
     CommonModule,
     ReactiveFormsModule,
-    ActivityModule,
     PortalModule,
     RouterModule
   ],
