+++ src/shared/components/event-start/index.ts
@@ -4,6 +4,6 @@ export * from './components/task-assigned/task-assigned.component';
 export * from './components/task-cancelled/task-cancelled.component';
 export * from './components/task-conflict/task-conflict.component';
 export * from './components/task-unassigned/task-unassigned.component';
-export * from './event-start.module';
 export * from './event-start.component';
+export * from './event-start.module';
 export * from './services';
